+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add creation of new os group `shinesolutions`
 - Add rsyslog configuration file for file/dir permissions
+- Introducing new configuration parameter to install AWS Kinesis agent
 
 
 ## 4.11.1 - 2019-09-12